refactor borrower book details to one Activity with different actions for book status
