+++ src/app/src/main/AndroidManifest.xml
@@ -14,22 +14,19 @@
 
         <activity android:name=".SignUpActivity" />
         <activity android:name=".ScanIsbnActivity" />
-        <activity android:name=".RequestedBookDetailsActivity" />
         <activity android:name=".PendingRequestsActivity" />
         <activity android:name=".MyProfileActivity" />
         <activity android:name=".MyBooksActivity" />
         <activity android:name=".MyBookDetailsActivity" />
         <activity android:name=".ManageRequestsActivity" />
-        <activity android:name=".ExploreBookDetailsActivity" />
         <activity android:name=".ExploreActivity" />
         <activity android:name=".EditProfileActivity" />
         <activity android:name=".EditBookActivity" />
-        <activity android:name=".BorrowedBookDetailsActivity" />
+        <activity android:name=".BorrowerBookDetailsActivity" />
         <activity android:name=".BorrowedActivity" />
         <activity android:name=".BorrowBookActivity" />
         <activity android:name=".AddBookActivity" />
         <activity android:name=".AcceptRequestsActivity" />
-        <activity android:name=".AcceptedBookDetailsActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
