+++ src/app/src/main/java/com/example/bookmark/PendingRequestsActivity.java
@@ -12,7 +12,7 @@ import com.example.bookmark.util.DialogUtil;
  * This activity shows a user a list of books that they have pending requests
  * for (books that they have requested that have a status of either REQUESTED
  * or ACCEPTED - they can select a book which takes them to the
- * RequestedBookDetailsActivity or the AcceptedBookDetailsActivity,
+ * BorrowerBookDetailsActivity or the BorrowerBookDetailsActivity,
  * respectively where they can see the books details.
  *
  * @author Ryan Kortbeek.
@@ -102,11 +102,6 @@ public class PendingRequestsActivity extends ListingBooksActivity {
      */
     @Override
     protected Class<?> getIntentDestination() {
-        if (selectedBook.getStatus() == Book.Status.ACCEPTED) {
-            return AcceptedBookDetailsActivity.class;
-        } else {
-            // Happens when selectedBook.getStatus() == Book.Status.REQUESTED
-            return RequestedBookDetailsActivity.class;
-        }
+        return BorrowerBookDetailsActivity.class;
     }
 }
