+++ src/app/src/main/res/layout/activity_borrower_book_details.xml
@@ -4,42 +4,40 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AcceptedBookDetailsActivity">
+    tools:context=".BorrowerBookDetailsActivity">
 
     <Button
         style="@style/Widget.MaterialComponents.Button"
-        android:id="@+id/accepted_details_action_btn"
+        android:id="@+id/borrower_book_details_action_btn"
         android:layout_width="219dp"
         android:layout_height="50dp"
         android:layout_marginTop="16dp"
-        android:onClick="handleBorrowButtonClick"
-        android:text="@string/borrow_book"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_book_status_text" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_book_status_text" />
 
     <ImageView
-        android:id="@+id/accepted_details_book_image"
+        android:id="@+id/borrower_book_details_book_image"
         android:layout_width="150dp"
         android:layout_height="150dp"
         android:layout_marginTop="24dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_description_text"
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_description_text"
         tools:srcCompat="@tools:sample/avatars" />
 
     <TextView
-        android:id="@+id/accepted_details_isbn_text"
+        android:id="@+id/borrower_book_details_isbn_text"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
         android:layout_marginTop="8dp"
         android:textSize="18sp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_author_text" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_author_text" />
 
     <TextView
-        android:id="@+id/accepted_details_description_text"
+        android:id="@+id/borrower_book_details_description_text"
         android:layout_width="0dp"
         android:layout_height="150dp"
         android:layout_marginStart="24dp"
@@ -47,10 +45,10 @@
         android:layout_marginEnd="24dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_isbn_text" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_isbn_text" />
 
     <TextView
-        android:id="@+id/accepted_details_title_text"
+        android:id="@+id/borrower_book_details_title_text"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
@@ -60,27 +58,27 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/accepted_details_owned_by"
+        android:id="@+id/borrower_book_details_owned_by"
         android:layout_width="219dp"
         android:layout_height="30dp"
         android:layout_marginTop="16dp"
         android:textAlignment="center"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_book_image" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_book_image" />
 
     <TextView
-        android:id="@+id/accepted_details_book_status_text"
+        android:id="@+id/borrower_book_details_book_status_text"
         android:layout_width="219dp"
         android:layout_height="30dp"
         android:layout_marginTop="16dp"
         android:textAlignment="center"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_owned_by" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_owned_by" />
 
     <TextView
-        android:id="@+id/accepted_details_author_text"
+        android:id="@+id/borrower_book_details_author_text"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
@@ -88,6 +86,6 @@
         android:textColor="#787878"
         android:textSize="18sp"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/accepted_details_title_text" />
+        app:layout_constraintTop_toBottomOf="@+id/borrower_book_details_title_text" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
